requirements.txt
Flask==3.0.3
APScheduler==3.10.4
requests==2.32.3
python-dotenv==1.1.1
Flask-SocketIO==5.5.1
eventlet==0.40.3
